Hello, Empire.
This commit is very much a work in progress; in fact, it's probably not working as of now.
It's my attempt to get https://github.com/remind101/empire to run on our infrastructure,
alongside all of the instance-backed apps we run in our stack.

I'm just happy to get to the point I could commit something.  Took a while to get here.